cdk/python/requirements.txt
aws-cdk-lib==2.92.0
constructs>=10.0.0,<11.0.0
pylint>=2.17.4
checkov>=2.3.318
